docker_apps/api_address_app/requirements.txt
annotated-types==0.5.0
dadata==21.10.1
anyio==3.7.1
click==8.1.6
colorama==0.4.6
fastapi==0.101.0
h11==0.14.0
idna==3.4
pydantic==2.1.1
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
gunicorn
mysqlclient==2.2.0
python-decouple==3.8
ordereddict==1.1
dynaconf